requirements.txt
wrapt==1.10.11
astroid==1.6.5
autopep8==1.4
CacheControl==0.12.5
cachetools==3.1.1
certifi==2019.6.16
chardet==3.0.4
chromedriver-binary==76.0.3809.25.0
Click==7.0
firebase-admin==2.17.0
flake8==3.6.0
Flask==1.1.1
Flask-Cors==3.0.8
google-api-core==1.13.0
google-api-python-client==1.7.9
google-auth==1.6.3
google-auth-httplib2==0.0.3
google-cloud-core==1.0.2
google-cloud-firestore==1.3.0
google-cloud-storage==1.16.1
google-resumable-media==0.3.2
googleapis-common-protos==1.6.0
grpcio==1.22.0
gunicorn==19.9.0
httplib2==0.13.0
idna==2.8
isort==4.3.4
itsdangerous==1.1.0
Jinja2==2.10.1
lazy-object-proxy==1.3.1
MarkupSafe==1.1.1
mccabe==0.6.1
msgpack==0.6.1
numpy==1.16.4
opencv-python==4.1.0.25
protobuf==3.9.0
pyasn1==0.4.5
pyasn1-modules==0.2.5
pycodestyle==2.4.0
pyflakes==2.0.0
pylint==1.9.3
pytz==2019.1
requests==2.22.0
rsa==4.0
selenium==3.141.0
six==1.11.0
uritemplate==3.0.0
urllib3==1.25.3
Werkzeug==0.15.4
wrapt==1.10.11
